requirements.txt
django==4.2.2
requests==2.32.3
python-dotenv==1.1.0
psycopg2-binary==2.9.10